requirements.txt
prometheus_client
psutil
pyyaml